+++ src/components/common/button/CancelWhiteBtn.tsx
@@ -1,4 +1,3 @@
-import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
 import Icons from '@/assets/svg/index';
@@ -13,7 +12,7 @@ function DeleteBtn() {
 
 export default DeleteBtn;
 
-const DeleteBtnCss = css`
+const DeleteBtnLayout = styled.button`
 	display: flex;
 	align-items: center;
 	justify-content: center;
@@ -21,10 +20,6 @@ const DeleteBtnCss = css`
 	height: 3.2rem;
 
 	border-radius: 10px;
-`;
-
-const DeleteBtnLayout = styled.button`
-	${DeleteBtnCss}
 `;
 
 const StlyedDeleteIcon = styled(Icons.DeleteIcon)`
