+++ src/pages/Setting.tsx
@@ -5,15 +5,15 @@ import EnterBtn from '@/components/common/button/EnterBtn';
 import OkayCancelBtn from '@/components/common/button/OkayCancelBtn';
 import ProgressBtn from '@/components/common/button/ProgressBtn';
 import RefreshBtn from '@/components/common/button/RefreshBtn';
-import SettingCheck2 from '@/components/common/button/SettingCheck2Btn';
-import SettingCheck3 from '@/components/common/button/SettingCheck3Btn';
-import SettingDeleteBtn from '@/components/common/button/SettingDeleteBtn';
-import SettingXBtn from '@/components/common/button/SettingXBtn';
+import SettingCheck2 from '@/components/common/button/settingBtn/SettingCheck2Btn';
+import SettingCheck3 from '@/components/common/button/settingBtn/SettingCheck3Btn';
+import SettingDeleteBtn from '@/components/common/button/settingBtn/SettingDeleteBtn';
+import SettingXBtn from '@/components/common/button/settingBtn/SettingXBtn';
 import SortBtn from '@/components/common/button/SortBtn';
-import StatusDoneBtn from '@/components/common/button/StatusDoneBtn';
-import StatusInProgressBtn from '@/components/common/button/StatusInProgressBtn';
-import StatusStagingBtn from '@/components/common/button/StatusStagingBtn';
-import StatusTodoBtn from '@/components/common/button/StatusTodoBtn';
+import StatusDoneBtn from '@/components/common/button/statusBtn/StatusDoneBtn';
+import StatusInProgressBtn from '@/components/common/button/statusBtn/StatusInProgressBtn';
+import StatusStagingBtn from '@/components/common/button/statusBtn/StatusStagingBtn';
+import StatusTodoBtn from '@/components/common/button/statusBtn/StatusTodoBtn';
 import TextBtn from '@/components/common/button/textBtn/TextBtn';
 import TimelineDeleteBtn from '@/components/common/button/TimelineDeleteBtn';
 import TodayPlusBtn from '@/components/common/button/TodayPlusBtn';
